pooled saom, parallel processing
pooled saom can now be fitted. implemented the RSiena fit on multiple cores. some minor changes in the SAOM data format.
changed the lags: not last three years anymore but now just a combined variable which tracks if there was any trade in the last 3 years. for arms and conventional trade